Using ffmpegthumbnailer

+++ scripts/generateFileData.sh
@@ -52,11 +52,14 @@ function createThumbnail() {
     one_third="$(( $format_duration / 3 + $random_offset ))"
     mkdir -p "$dataDir"
     #ffmpeg -i "$file" -loglevel quiet -ss $one_third -f image2 -vframes 1 "$dataDir/preview.png" 2>/dev/null
-    cd "$dataDir"
-    mplayer --quiet "$( readlink -f "$file" )" -vo png:z=9 -ao null -frames 1 -ss "$one_third" 2>&1 > /dev/null
-    mv 00000001.png preview.png
-    rm 00000*.png
-    cd -
+    ffmpegthumbnailer -i "$file" -o "$dataDir/preview.png" -s 0 -t $one_third -f
+    if ! [ -r "$dataDir/preview.png" ]; then
+	    cd "$dataDir"
+	    mplayer --quiet "$( readlink -f "$file" )" -vo png:z=9 -ao null -frames 1 -ss "$one_third" 2>&1 > /dev/null
+	    mv 00000001.png preview.png
+	    rm 00000*.png
+	    cd -
+    fi
     if ! [ -r "$dataDir/preview.png" ]; then
 	set -x
         ffmpeg -i "$( readlink -f "$file")" -loglevel error -ss $one_third -f image2 -vframes 1 "$dataDir/preview.png"
